requirements-agent.txt
llama-cpp-python>=0.2.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pytest
